Setting scope and name in Subscriber page...
But I think this page does not have to exist. The "create" page is the page each participant will actually go through.

Every participant is a Publisher, that in turn subscribers to N-number of streams during an event. The are a Publisher in the sense that they are publishing there local stream to be subscribed to by other
publishers.